Added Payments
Worked on adding a "payment" function for the tickets while also calculating the total cost of tickets based on seat and package selections. Struggling with the back button, but will look to maybe implement in a more advance creation of this. Re-wrote a lot of code to then have to rewrite it to its original form. Next step is to add the receipt function and then the admin file with the sales implementation. After finishing those, I will look into either advancing this code or taking a break and learning some MATLAB.