Receipt Function Added
Went on to add my receipt function and added a yes/no selection function. One thing I want to go back and do later is allow for all my information to be stored in main as I realized I had to rewrite it a bunch. May add a couple new functions with it. I want to finish my admin setup before. Not too much left when it comes to getting the skeleton of this project done though.