Package Add Function
Just a quick little work on the package adding function. It successfully adds the package name and makes sure to check with admin that the name is correct before adding. Appears properly in customer view as well! Wasn't too much of work to add. Will continue on this project tomorrow or Sunday!