Saved Amount Added
Last thing to add for me was telling customers how much they saved from the sales event when it's going on. Will need to go back and make sure my sales_event function resets back to original prices if the admin decides to re-input the sales percentage again in the same session.