Updated back button
A little messy of a back button, but it currently works. It won't automatically prompt to modify another seat change if you go back in the seatings listed section. The back button works more properly when asked to input a number or go back. It will only move on if you actually put in a price that you confirm and then have update into the dictionary. Where I update the seating price, it could be condensed into one line, but I left it the way it is for readability