Seat Editor FInished
Expanded on my seat editor where admin can delete seats, add seats, or just change the name. Added a submenu for this. The back function works properly and so does everything else. Would want to make this look a lot cleaner when I'm more developed as a programmer, but satisfied with what I have currently. Will be trying to finish the package editing function and an admin view section before my semester starts up in the next three days!